Move query logic to a separate module, `QueryFacade`, that knows nothing about http Introduce `AuthFacade` that leverages `QueryFacade`
Change-type: patch
